Move Nav into own folder. move styles out
